requirements.txt
python-decouple
requests
openai
streamlit
fitz